Updated probe mapping files
Updated files from probe mapping.
Includes a filter list based only on source sequences. Probes listed
here are removed if there was no or multiple mappings to the source
genome. Furthermore a filter list based only on source and probe
sequences. Probes listed here are removed if there was no or multiple
mappings to the source genome or to the bisulfite treated genome.